Fix an issue where loaded items would not be correct
This also removes client-side paging, which is troublesome
and not a requirement anymore now that I know that the API
implements paging.

+++ _list/src/main/kotlin/list/impl/RefreshResponse.kt
@@ -4,10 +4,10 @@ import com.squareup.moshi.Json
 
 internal class RefreshResponse private constructor(
     @field:Json(name = "page")
-  val page: Long?,
+  val page: Int?,
     @field:Json(name = "results")
   val results: List<ListItem>,
     @field:Json(name = "total_results")
-  val totalResults: Long?,
+  val totalResults: Int?,
     @field:Json(name = "total_pages")
-  val totalPages: Long?)
+  val totalPages: Int?)
